moved BrandSearchQuery and all related code to the frontend-api and framework packages
