+++ packages/frontend-api/src/Resources/config/graphql-types/QueryDecorator.types.yaml
@@ -3,6 +3,13 @@ QueryDecorator:
     decorator: true
     config:
         fields:
+            brandSearch:
+                type: "[Brand!]!"
+                resolve: "@=query('brandSearchQuery', args)"
+                args:
+                    search:
+                        type: "String!"
+                description: "Returns list of searched brands"
             categories:
                 type: '[Category!]!'
                 resolve: "@=query('categoriesQuery')"
