--- project-base/app/config/graphql/types/Query/Product/Brand/BrandQuery.types.yaml
@@ -1,11 +1,0 @@
-BrandQuery:
-    type: object
-    config:
-        fields:
-            brandSearch:
-                type: "[Brand!]!"
-                resolve: "@=query('brandSearchQuery', args)"
-                args:
-                    search:
-                        type: "String!"
-                description: "Returns list of searched brands"
