+++ project-base/app/config/graphql/types/Query/Query.types.yaml
@@ -11,7 +11,6 @@ Query:
         - 'OrderQuery'
         - 'PersonalDataQuery'
         - 'ProductQuery'
-        - 'BrandQuery'
         - 'SlugQuery'
         - 'NotificationBarQuery'
         - 'CartQuery'
