+++ project-base/app/src/Model/Product/Filter/Elasticsearch/ProductFilterConfigFactory.php
@@ -17,8 +17,7 @@ use Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfigFactory as B
  * @property \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser
  * @property \App\Model\Product\Parameter\ParameterFacade $parameterFacade
  * @property \App\Model\Product\Flag\FlagFacade $flagFacade
- * @property \App\Model\Product\Brand\BrandFacade $brandFacade
- * @method __construct(\App\Model\Product\Filter\ParameterFilterChoiceRepository $parameterFilterChoiceRepository, \App\Model\Product\Filter\FlagFilterChoiceRepository $flagFilterChoiceRepository, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \App\Model\Product\Filter\BrandFilterChoiceRepository $brandFilterChoiceRepository, \App\Model\Product\Filter\PriceRangeRepository $priceRangeRepository, \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterElasticFacade $productFilterElasticFacade, \App\Model\Product\Parameter\ParameterFacade $parameterFacade, \App\Model\Product\Flag\FlagFacade $flagFacade, \App\Model\Product\Brand\BrandFacade $brandFacade)
+ * @method __construct(\App\Model\Product\Filter\ParameterFilterChoiceRepository $parameterFilterChoiceRepository, \App\Model\Product\Filter\FlagFilterChoiceRepository $flagFilterChoiceRepository, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \App\Model\Product\Filter\BrandFilterChoiceRepository $brandFilterChoiceRepository, \App\Model\Product\Filter\PriceRangeRepository $priceRangeRepository, \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterElasticFacade $productFilterElasticFacade, \App\Model\Product\Parameter\ParameterFacade $parameterFacade, \App\Model\Product\Flag\FlagFacade $flagFacade, \Shopsys\FrameworkBundle\Model\Product\Brand\BrandFacade $brandFacade)
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterConfig createForCategory(string $locale, \App\Model\Category\Category $category)
  * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[] getSortedParameterFilterChoicesForCategory(\Shopsys\FrameworkBundle\Model\Product\Filter\ParameterFilterChoice[] $aggregatedParameterFilterChoices, \App\Model\Category\Category $category)
  */
